use size_t for strlen

+++ ft_strlen.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	unsigned int	count;
+	size_t	count;
 
 	count = 0;
 	while (*(s++))
